fix: move yubikey_utils from tests folder
Use random generator suitable for security/cryptographic purposes (bandit).
